Remove channel based implementation
